Save check fix (#42)
This should remove the "unsaved changes" notification that appears for no reason sometimes.

* Automatic resave when chip version doesn't match

Instead of leaving the user with the responsability to save themself when there is simply a version difference, the chip will now save itself when opened if it is the case.

* Fixes save check regarding pin face

Fixes the pin face of the pins in the chipdescriptions generated when checking for modifications

---------

Co-authored-by: firecerne <105079722+firecerne@users.noreply.github.com>