Hotkey customization (#47)
* Shortcuts refactor (wtih new ShortcutSettings file)

Changes introduced to how the shortcuts / hotkeys work to enable customization in the future.
Hotkeys are now stored in a ShortcutSettings file (it has support for modifiers, alternative modifiers, alternative keys, and forbidden modifiers).

* Change shortcut preview to the one set in ShortcutSettings.

The shortcut preview string on the "menu" button on the bottom bar now directly corresponds to the one set in the ShortcutSettings (whatever it may be).

Added a MenuHelper method to return the string representation of a shortcut.

* Added automatic padding to bottom bar ui menu collapsable

* Custom Hotkeys edition menu

Added a menu to edit hotkeys in game. (in settings > controls)
Corrected a small error that made complex shortcuts (with alternative keys/modifiers) not register.
Changed Shortcuts from struct to class to make things easier.
Added methods to InputHelper to more precisely record inputs (useful for detecting hotkey setting inputs and other cases where it is needed)